+++ CHANGELOG.md
@@ -7,14 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [v0.1.0] - 2019-12-06
 ### Added
 - Add CHANGELOG
+- Add graveyard shortcut
 
 ### Changed
 - Update the dependences
 - Change readme to markdown
+- Change macos default trash directory to  `~/.Trash`
+- Minimum Rust supported channel > 1.39.0
 
-## Fix
-- Fix travis build
-
+### Fix
+- Update travis build
 
 [Unreleased]: https://github.com/alongwy/safe-rm/compare/v0.1.0...HEAD
 [v0.1.0]: https://github.com/alongwy/safe-rm/compare/fork...v0.1.0